修改返回类型

+++ compute-service/src/main/java/spring/cloud/mapper/ds1/JournalizingInfoMapper.java
@@ -1,8 +1,5 @@
 package spring.cloud.mapper.ds1;
 
-import java.util.List;
-import java.util.Map;
-
 import org.apache.ibatis.annotations.Param;
 
 public interface JournalizingInfoMapper {
@@ -11,13 +8,13 @@ public interface JournalizingInfoMapper {
 	 * 汇总科目金额
 	 * @return
 	 */
-	public Integer sumCurrentSubject(@Param("id") String id, @Param("companyId") String companyId, @Param("type") Short type, @Param("datasource") String datasource,@Param("startDate") String startDate,@Param("endDate") String endDate);
+	public double sumCurrentSubject(@Param("id") String id, @Param("companyId") String companyId, @Param("type") Short type, @Param("datasource") String datasource,@Param("startDate") String startDate,@Param("endDate") String endDate);
 	
 	/**
 	 * 汇总科目及其子科目金额 
 	 * @return
 	 */
-	public Integer sumContainChildrenSubject(@Param("id") String id, @Param("companyId") String companyId, @Param("type") Short type, @Param("datasource") String datasource,@Param("startDate") String startDate,@Param("endDate") String endDate);
+	public double sumContainChildrenSubject(@Param("id") String id, @Param("companyId") String companyId, @Param("type") Short type, @Param("datasource") String datasource,@Param("startDate") String startDate,@Param("endDate") String endDate);
 	
 	public void insertApiHistoryRecord(@Param("ip") String ip);
